refactor: datagen everything
This also fixes recipes that didn't appear in the recipe book, such as
module upgrades. There's still a few small hiccups (for example if you
have all the requirements for a laser, it won't automatically move the
book into the recipe slot), but most of the recipes are working now.