fix: lasers mimic vanilla breaking behavior
We now mimic exactly the vanilla block breaking behavior according to
ServerPlayerInteractionManager. This means that everything that should
drop blocks will drop a block, including edge cases such as computers
and shulkers, who spawn items in onBreak() in certain situations.

As a result of performing the vanilla checks, this now also means that
firing lasers by hand in creative mode no longer drops items (except for
the aforementioned edge cases). It will work exactly as if you had
left-clicked the block. This might not be the most desirable behavior,
but it's better than hunting down all the individual edge cases to
prevent creative-mode-only dupes. Fixing this (perhaps with a tag?) is
potential future work.

Fixes #24