fix: allow stacking in armor slots
This is inconsistent with vanilla behavior (you cannot stack
jack-o-lanterns or skulls in your helmet slot), but unfortunately
isValid() is only ever passed a copy stack of size 1. This means you'll
still always be able to insert a stack of size >1. So instead, we opt
for the more self-consistent behavior of allowing you to add to an
existing stack, if the item is stackable.

Co-authored-by: Alyssa May <may@alyss.am>