fix: Item meta proxies/wrappers now contain all metadata
Previously MetaWrapper used Plethora for fetching item metadata. This is
correct(TM) (it ensures the context is used), however it also means that
all the built-in CC:T item details are missing, making the returned
table a little useless.

This removes MetaWrapper, and replaces it with a DetailsMetaWrapper
class, which has a similar idea (/just/ expose metadata and no other
methods), but goes through a DetailsRegistry instead.